refactor: use cookies insted of localstorage
