+++ package.json
@@ -32,6 +32,7 @@
     "@types/jest": "27.4.0",
     "@types/js-cookie": "3.0.2",
     "@types/jsdom": "16.2.14",
+    "@types/qs": "6.9.7",
     "@typescript-eslint/eslint-plugin": "5.9.0",
     "@typescript-eslint/parser": "5.9.0",
     "eslint": "8.6.0",
@@ -46,6 +47,7 @@
   },
   "packageManager": "yarn@3.2.0-rc.11",
   "dependencies": {
-    "js-cookie": "3.0.1"
+    "js-cookie": "3.0.1",
+    "qs": "6.10.3"
   }
 }
