+++ src/utils/cookie.ts
@@ -1,8 +1,15 @@
 import Cookies from 'js-cookie';
 
+interface Session {
+  id: string;
+  token: string;
+}
+
 export const COOKIE_KEYS = {
   ACCEPT_COOKIES_KEY: 'accept-all-cookies',
   SESSION_KEY: 'session',
+  STORED_SESSIONS_KEY: 'storedSessions',
+  REDIRECT_URL_KEY: 'redirectUrl',
 };
 
 export const hasAcceptedCookies = () =>
@@ -11,11 +18,34 @@ export const hasAcceptedCookies = () =>
 /**
  * @param  {string|null} token token to set in the session key. A null value will remove it.
  */
-export const setSessionCookie = (token: string | null, domain: string) => {
+export const setCurrentSession = (token: string | null, domain: string) => {
   if (!token) {
     return Cookies.remove(COOKIE_KEYS.SESSION_KEY);
   }
-  return Cookies.set(COOKIE_KEYS.SESSION_KEY, token, { domain });
+  return Cookies.set(COOKIE_KEYS.SESSION_KEY, token, { domain, secure: true });
 };
 
-export const getSessionCookie = () => Cookies.get(COOKIE_KEYS.SESSION_KEY);
+export const getCurrentSession = () => Cookies.get(COOKIE_KEYS.SESSION_KEY);
+
+export const getStoredSessions = (): Session[] =>
+  JSON.parse(Cookies.get(COOKIE_KEYS.STORED_SESSIONS_KEY) ?? '[]');
+
+export const addStoredSession = (payload: Session, domain: string) => {
+  const sessions = getStoredSessions();
+
+  // add session if doesn't exist
+  if (!sessions.find(({ id }) => id === payload.id)) {
+    Cookies.set(
+      COOKIE_KEYS.STORED_SESSIONS_KEY,
+      JSON.stringify(sessions.concat([payload])),
+      { domain, secure: true },
+    );
+  }
+};
+
+export const saveUrlForRedirection = (link: string, domain: string) => {
+  Cookies.set(COOKIE_KEYS.REDIRECT_URL_KEY, link, { domain, secure: true });
+};
+
+export const getUrlForRedirection = () =>
+  Cookies.get(COOKIE_KEYS.REDIRECT_URL_KEY);
