+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './navigation';
-export * from './localStorage';
 export * from './cookie';
