+++ src/utils/navigation.ts
@@ -1,4 +1,5 @@
-import { getUrlForRedirection } from './localStorage';
+import qs from 'qs';
+import { getUrlForRedirection } from './cookie';
 
 interface RedirectOptions {
   openInNewTab?: boolean;
@@ -29,3 +30,6 @@ export const redirectToSavedUrl = (
 
   return false;
 };
+
+export const redirectToSignIn = ({ host, to }: { host: string; to?: string }) =>
+  `${host}/signin${qs.stringify({ to }, { addQueryPrefix: true })}`;
